# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutors/{slug}:
    get:
      operationId: getTutor
      tags: [tutors]
      summary: Get a tutor profile
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Tutor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutor"
        "404":
          $ref: "#/components/responses/NotFound"
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List tutor-written resources
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: query
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resource summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
  /resources/{slug}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a full tutor-written resource
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Tutor:
      type: object
      required: [id, slug, name, firstName, lastName, initials, subject, support, profileSummary, university, qualification, bio, style, teachingIntro, teachingPoints, rate, availability, tint, resources]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        initials: { type: string }
        subject: { type: string }
        support: { type: string }
        profileSummary: { type: string, maxLength: 200 }
        university: { type: string }
        qualification: { type: string }
        bio: { type: string }
        style: { type: string }
        teachingIntro: { type: string }
        teachingPoints:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TeachingPoint"
        rate: { type: number }
        availability: { type: string, enum: [accepting, limited, unavailable] }
        tint: { type: string }
        resources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
...
    Resource:
      type: object
      required: [id, slug, title, tutorSlug, tutorName, tutorTint, subject, level, type, readMinutes, excerpt, publishedAt, tint]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        tutorSlug: { type: string }
        tutorName: { type: string }
        tutorTint: { type: string }
        subject: { type: string }
        level: { type: string }
        type: { $ref: "#/components/schemas/ResourceType" }
        readMinutes: { type: integer }
        excerpt: { type: string }
        publishedAt: { type: string, format: date }
        tint: { type: string }
    ResourceDetail:
      allOf:
        - $ref: "#/components/schemas/Resource"
        - type: object
          required: [body, sections, related]
          properties:
            body: { type: string }
            sections:
              type: array
              items:
                $ref: "#/components/schemas/ResourceSection"
            related:
              type: array
              items:
                $ref: "#/components/schemas/Resource"
    ResourceType:
      type: string
      enum: [Study note, Guide, Essay, Revision notes]
...
    ResourceSection:
      type: object
      required: [id, heading, body]
      properties:
        id: { type: string }
        heading: { type: string }
        body: { type: string }
    TeachingPoint:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 2 }
        body: { type: string, minLength: 10 }
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }