# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from './resourceType';

export interface Resource {
  id: number;
  slug: string;
  title: string;
  tutorSlug: string;
  tutorName: string;
  tutorTint: string;
  subject: string;
  level: string;
  type: ResourceType;
  readMinutes: number;
  excerpt: string;
  publishedAt: Date;
  tint: string;
}