# lib/api-zod/src/generated/types/resourceDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from './resource';
import type { ResourceSection } from './resourceSection';

export type ResourceDetail = Resource & {
  body: string;
  sections: ResourceSection[];
  related: Resource[];
};