# lib/api-zod/src/generated/types/resourceSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceSection {
  id: string;
  heading: string;
  body: string;
}