# lib/api-zod/src/generated/types/resourceType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ResourceType = typeof ResourceType[keyof typeof ResourceType];


export const ResourceType = {
  Study_note: 'Study note',
  Guide: 'Guide',
  Essay: 'Essay',
  Revision_notes: 'Revision notes',
} as const;